+++ Module/Public/Core/Program.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <Core/Server.h>
+#include <Core/Client.h>
 
 #include <SDK/TypeInfo.h>
 #include <SDK/SDK.h>
@@ -14,21 +15,18 @@
 
 #include <Entity/NativeEntityManager.h>
 #include <Script/ScriptManager.h>
-#include <Voip/VoipManager.h>
 
 #include <Windows.h>
 #include <condition_variable>
 
 #define OFFSET_GLOBAL_CLIENT 0x143DCB9D0
 #define OFFSET_GLOBAL_SETTINGS_MANAGER 0x143D11950
-#define OFFSET_GET_CLIENT_INSTANCE 0x14659DE50
 
 namespace Kyber
 {
 TL_DECLARE_FUNC(0x1401F7BD0, DataContainer*, SettingsManager_getSettingsObject, void* inst, const char* identifier);
 TL_DECLARE_FUNC(0x1401F83F0, void, MessageManager_queueMessage, void* pMessageManagerImpl, Message* pMessage, float delayTime);
 
-__int64 ClientStateChangeHk(__int64 a1, ClientState currentClientState, ClientState lastClientState);
 const char* GetHostIdHk(__int64 inst);
 void MessageManagerDispatchMessageHk(void* pMessageManagerImpl, Message* pMessage);
 void OnGotDisconnectedHk(void* inst, SecureReason reason, const eastl::string& reasonText);
@@ -53,14 +51,6 @@ public:
     void InitializeGamePatches();
     void Initialize();
 
-    void RegisterClientUpdatePassListener(ClientUpdatePassListener* listener);
-    void HandleClientServerJoin(NetworkCreatePlayerMessage* message);
-    
-    // If proxied is true, IP is assumed to be a proxy ID
-    void JoinServer(const std::string& id, std::string ip, uint16_t port, bool spectate, bool proxied = false, bool changeState = true);
-
-    void AttemptJoinVoip();
-
     void* GetSettingsManager()
     {
         return *reinterpret_cast<void**>(OFFSET_GLOBAL_SETTINGS_MANAGER);
@@ -70,13 +60,6 @@ public:
     T* GetSettingsObject(const char* identifier)
     {
         return reinterpret_cast<T*>(SettingsManager_getSettingsObject(GetSettingsManager(), identifier));
-    }
-
-    __int64 ChangeClientState(ClientState currentClientState)
-    {
-        return ClientStateChangeHk(
-            *reinterpret_cast<__int64*>(*reinterpret_cast<__int64*>(((__int64 (*)(void))OFFSET_GET_CLIENT_INSTANCE)() + 0x20) + 0x28),
-            currentClientState, m_clientState);
     }
 
     API* GetAPI()
@@ -90,30 +73,20 @@ public:
     ModData m_modData;
 
     Server* m_server;
+    Client* m_client;
     Console* m_console;
 
     EntityManager* m_entityManager;
     ScriptManager* m_scriptManager;
 
     KyberSettingsManager* m_settingsManager;
-    VoipManager* m_voipManager;
-
-    SocketManager* m_clientSocketManager;
-
-    ClientState m_clientState;
 
     std::vector<std::function<void()>> m_consoleRegistrationCallbacks;
-    std::vector<ClientUpdatePassListener*> m_clientUpdatePassListeners;
 
     std::mutex m_startupMutex;
     std::condition_variable m_startupCondition;
 
-    std::string m_joinToken;
-
     bool m_startupInitialized;
-    bool m_joining;
-    bool m_spectator;
-    bool m_connected;
     bool m_allowInteraction;
     bool m_isDedicatedServer;
     bool m_messageDebugEnabled;
