+++ Module/Source/Network/UDPSocket.cpp
@@ -239,7 +239,7 @@ bool UDPSocket::Listen(const SocketAddr& address, bool blocking)
     if (m_info.isProxied && m_direction == ProtocolDirection::Serverbound)
     {
         m_sockets.push_back(WebSocket("client", 0, m_proxyQueue));
-        m_sockets.back().ConnectAsClient(m_info.proxyAddress, g_program->m_joinToken);
+        m_sockets.back().ConnectAsClient(m_info.proxyAddress, g_program->m_client->m_joinToken);
     }
     else if (m_direction == ProtocolDirection::Clientbound && g_program->m_server->IsRunning() && g_program->m_server->m_onlineMode)
     {
@@ -247,7 +247,7 @@ bool UDPSocket::Listen(const SocketAddr& address, bool blocking)
         if (pinnedProxy != nullptr)
         {
             m_sockets.push_back(WebSocket("pinned", 0, m_proxyQueue));
-            m_sockets.back().ConnectAsServer(pinnedProxy, g_program->m_joinToken);
+            m_sockets.back().ConnectAsServer(pinnedProxy, g_program->m_client->m_joinToken);
         }
         else
         {
@@ -258,7 +258,7 @@ bool UDPSocket::Listen(const SocketAddr& address, bool blocking)
             for (const auto& proxy : proxies)
             {
                 m_sockets.push_back(WebSocket(proxy.id(), id++, m_proxyQueue));
-                m_sockets.back().ConnectAsServer(proxy.ip(), g_program->m_joinToken);
+                m_sockets.back().ConnectAsServer(proxy.ip(), g_program->m_client->m_joinToken);
             }
         }
     }
